Modularised File Structure
updated readme, and modularised file structure.
