+++ training_history.py
@@ -1,0 +1,24 @@
+import streamlit as st
+import matplotlib.pyplot as plt
+
+def plot_training_history(history):
+    st.write("Training and Validation Metrics:")
+    acc = history.history['accuracy']
+    val_acc = history.history['val_accuracy']
+    loss = history.history['loss']
+    val_loss = history.history['val_loss']
+
+    epochs_range = range(len(acc))
+
+    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 4))
+    ax1.plot(epochs_range, acc, label='Training Accuracy')
+    ax1.plot(epochs_range, val_acc, label='Validation Accuracy')
+    ax1.set_title('Training and Validation Accuracy')
+    ax1.legend()
+
+    ax2.plot(epochs_range, loss, label='Training Loss')
+    ax2.plot(epochs_range, val_loss, label='Validation Loss')
+    ax2.set_title('Training and Validation Loss')
+    ax2.legend()
+
+    st.pyplot(fig)
